requirements.txt
fastapi
uvicorn[standard]
asyncpg
psycopg2-binary
pgvector
sqlalchemy[asyncio]
groq
httpx
python-dotenv
pydantic
sentence-transformers
sse-starlette
